backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
python-socketio==5.11.2
redis==5.0.4
httpx==0.27.0
authlib==1.3.1
email-validator==2.1.1
